+++ src/main/java/com/simibubi/create/content/redstone/contact/ContactMovementBehaviour.java
@@ -41,7 +41,7 @@ public class ContactMovementBehaviour implements MovementBehaviour {
 		Vec3 contact = Vec3.atLowerCornerOf(block.getValue(RedstoneContactBlock.FACING)
 			.getUnitVec3i());
 		contact = context.rotation.apply(contact);
-		Direction direction = Direction.getNearest(contact.x, contact.y, contact.z, Direction.UP);
+		Direction direction = Direction.getNearest((int) contact.x, (int) contact.y, (int) contact.z, Direction.UP);
 
 		if (visitedState.getValue(RedstoneContactBlock.FACING) != direction.getOpposite())
 			return;
